requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mssql==1.8
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pytds==1.5
django-sqlserver==1.11
djangorestframework==3.12.2
Markdown==3.3.3
mysqlclient==2.0.1
Pillow==8.0.1
pyodbc==4.0.30
pypiwin32==223
python-tds==1.10.0
pytz==2020.4
pywin32==300
six==1.15.0
sqlparse==0.4.1
